git-svn-id: svn://120.24.217.103/repository/outsource_projects/wu_xb/android/shop/XiaoeShop@253375 d1851369-bd24-4d56-8f42-d5f936fd0249

+++ app/src/main/java/xiaoe/com/shop/adapter/download/DownloadProceedChildListHolder.java
@@ -65,7 +65,7 @@ public class DownloadProceedChildListHolder extends BaseViewHolder implements Vi
             downloadStartIcon.setImageResource(R.mipmap.download_play);
         }
         float progress = info.getProgress() / (1024f * 1024f);
-        long totalSize = info.getTotalSize() / (1024 * 1024);
+        float totalSize = info.getTotalSize() / (1024f * 1024f);
         float strCount = (float)(Math.round(progress*10*2))/(10*2);
         float strTotal = (float)(Math.round(totalSize*10*2))/(10*2);
         downloadSpeed.setText(strCount+"MB/"+strTotal+"MB");
